backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
packaging==23.2
pdf2image==1.17.0
pillow==10.2.0
pytesseract==0.3.10
sqlparse==0.4.4
